Убрал некоторые классы
