+++ Client/src/main/java/clientapp/MainOutboundHandler.java
@@ -1,12 +1,7 @@
 package clientapp;
 
-import common.MyMessage;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelOutboundHandlerAdapter;
-import io.netty.channel.ChannelPromise;
-
-import java.net.SocketAddress;
-import java.util.Scanner;
 
 public class MainOutboundHandler extends ChannelOutboundHandlerAdapter {
