+++ Common/src/main/java/common/MyFileSend.java
@@ -1,0 +1,40 @@
+package common;
+
+import io.netty.buffer.ByteBuf;
+import io.netty.buffer.ByteBufAllocator;
+import io.netty.channel.*;
+
+import java.io.*;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Path;
+
+public class MyFileSend {
+
+    public static void sendFile(Path path, Channel channel, ChannelFutureListener finishListener) throws IOException {
+        FileRegion region = new DefaultFileRegion(path.toFile(), 0, Files.size(path));
+
+        ByteBuf buf = null;
+        buf = ByteBufAllocator.DEFAULT.directBuffer(1);
+        buf.writeByte((byte) 25);
+        channel.write(buf);
+
+        byte[] filenameBytes = path.getFileName().toString().getBytes(StandardCharsets.UTF_8);
+        buf = ByteBufAllocator.DEFAULT.directBuffer(4);
+        buf.writeInt(filenameBytes.length);
+        channel.write(buf);
+
+        buf = ByteBufAllocator.DEFAULT.directBuffer(filenameBytes.length);
+        buf.writeBytes(filenameBytes);
+        channel.write(buf);
+
+        buf = ByteBufAllocator.DEFAULT.directBuffer(8);
+        buf.writeLong(Files.size(path));
+        channel.write(buf);
+
+        ChannelFuture transferOperationFuture = channel.writeAndFlush(region);
+        if (finishListener != null) {
+            transferOperationFuture.addListener(finishListener);
+        }
+    }
+}
