+++ Server/src/main/java/serverapp/Authorization.java
@@ -1,0 +1,4 @@
+package serverapp;
+
+public class Authorization {
+}
