+++ Server/src/main/java/serverapp/MainServerHandler.java
@@ -1,7 +1,7 @@
 package serverapp;
 
-import common.MyFile;
-import common.MyMessage;
+import common.MyFileReceive;
+import common.MyFileSend;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import io.netty.channel.ChannelHandlerContext;
@@ -11,8 +11,6 @@ import io.netty.util.CharsetUtil;
 public class MainServerHandler extends ChannelInboundHandlerAdapter {
 
     private String nickName;
-    private MyMessage myMessage = new MyMessage();
-    private MyFile myFile = new MyFile();
     private String str;
     private String prev = "";
 
@@ -32,22 +30,18 @@ public class MainServerHandler extends ChannelInboundHandlerAdapter {
         ByteBuf in = (ByteBuf) msg;
         // обрабатываем первую отправку чтобы узнать наличие /file в строке
         if (!prev.equals("/file")) {
-            str = myMessage.formReceiveMsg(in);
+            str = in.toString(CharsetUtil.UTF_8);
+            System.out.println(nickName + " написал: " + str);
         }
         // т.к предидущая строка содержала /file то обрабатываем прием файла, после чего скидываем предидущий показатель на ""
         if(prev.equals("/file")) {
-            myFile.reciveFile(in);
+            MyFileReceive.reciveFile(in);
             prev = "";
         }
         // после получении строки проверяем начинается ли она с /file
         if (str.startsWith("/file")) {
             // помечаем что предидущая строка содержала /file
             prev = "/file";
-//            ctx.pipeline().addLast(new MainFileServerHandler(in));
-//            ctx.pipeline().remove(this);
-
-        } else {
-            System.out.println(nickName + " написал: " + str);
         }
     }
